moved classes to combined nmap/amass scan
