+++ scb-scanprocesses/combined-amass-nmap-process/src/main/java/io/securecodebox/scanprocesses/amassnmap/util/HttpHeaders.java
@@ -1,4 +1,4 @@
-package io.securecodebox.scanprocess.nmap.util;
+package io.securecodebox.scanprocesses.amassnmap.util;
 
 import io.securecodebox.model.findings.Finding;
